updated useEffect in every component, also refactored code in reservations component
